requirements.txt
# Large Language Models in Cybersecurity - Requirements
# Python 3.8+

# OpenAI API (GPT-5, GPT-5.1)
openai>=1.12.0

# Anthropic Claude API (optional)
anthropic>=0.18.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Web requests
requests>=2.31.0

# JSON handling
jsonschema>=4.17.0

# CLI tools
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security tools integration (optional)
python-nmap>=0.7.1
scapy>=2.5.0

# Type hints
typing-extensions>=4.7.0
